requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
dj-database-url==2.3.0
Django==5.1.6
djangorestframework==3.15.2
dotenv==0.9.9
gunicorn==23.0.0
idna==3.10
packaging==24.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
